requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
httpx==0.25.1
selectolax==0.3.17
feedparser==6.0.10
openai==1.3.5
numpy==1.24.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
redis==5.0.1
tenacity==8.2.3
python-dotenv==1.0.0
apscheduler==3.10.4
jinja2==3.1.2

# Playwright MCP dependencies (optional - for authenticated scrapers)
aiohttp==3.10.10
sseclient-py==1.8.0